Remove PotterhatOrchestrator
