+++ apps/potterhat_node/config/config.exs
@@ -1,6 +1,8 @@
 use Mix.Config
 
 config :potterhat_node,
+  nodes: {:apply, {PotterhatNode.EnvConfigProvider, :get_configs, []}},
+  active_nodes: [],
   retry_period_ms: 5000,
   listen_events: [:new_head]
