+++ apps/potterhat_node/config/test.exs
@@ -1,1 +1,4 @@
 use Mix.Config
+
+# Sets nodes to an empty list as the tests don't need the nodes that are set in the env vars.
+config :potterhat_node, nodes: []
