--- apps/potterhat_orchestrator/README.md
@@ -1,3 +1,0 @@
-# PotterhatOrchestrator
-
-Orchestrates various app components to form a single resilient Ethereum client.
