--- apps/potterhat_orchestrator/config/dev.exs
@@ -1,1 +1,0 @@
-use Mix.Config
