--- apps/potterhat_orchestrator/config/prod.exs
@@ -1,1 +1,0 @@
-use Mix.Config
