--- apps/potterhat_orchestrator/config/test.exs
@@ -1,4 +1,0 @@
-use Mix.Config
-
-# Sets nodes to an empty list as the tests don't need the nodes that are set in the env vars.
-config :potterhat_orchestrator, nodes: []
