--- apps/potterhat_orchestrator/mix.exs
@@ -1,35 +1,0 @@
-defmodule PotterhatOrchestrator.MixProject do
-  use Mix.Project
-
-  def project do
-    [
-      app: :potterhat_orchestrator,
-      version: "0.1.0",
-      build_path: "../../_build",
-      config_path: "../../config/config.exs",
-      deps_path: "../../deps",
-      lockfile: "../../mix.lock",
-      elixir: "~> 1.8",
-      elixirc_paths: elixirc_paths(Mix.env),
-      start_permanent: Mix.env() == :prod,
-      deps: deps()
-    ]
-  end
-
-  def application do
-    [
-      extra_applications: [:logger],
-      mod: {PotterhatOrchestrator.Application, []}
-    ]
-  end
-
-  # Specifies which paths to compile per environment.
-  defp elixirc_paths(:test), do: ["lib", "test/support"]
-  defp elixirc_paths(_),     do: ["lib"]
-
-  defp deps do
-    [
-      {:potterhat_node, in_umbrella: true}
-    ]
-  end
-end
